// src/app/modules/admin/models/tableUser.ts
export class TableUser {
    email: string;
    fechaRegistroStr: string;
    lastTimeLoginDate: string;
    repLocalProvCode: string;
    repLocalProvInfo: string;
    roleCode: string;
    roleInfo: string;
    userName: string;
    userCode: string;
    userId: number;
    roleId: number;
    repLocalProvId: number;
    userSupervisorCode: string;
}